Relocating the configuration related functionalities into the config.c and config.h files
