+++ privatebet/bet.h
@@ -176,7 +176,4 @@ struct pair256 p2p_bvv_init(bits256 *keys, struct pair256 b_key, bits256 *blindi
 			    bits256 deckid);
 
 bits256 curve25519_fieldelement(bits256 hash);
-cJSON *bet_read_json_file(char *file_name);
-void bet_parse_config_file();
-void bet_parse_notary_file();
 #endif
