+++ privatebet/config.h
@@ -1,0 +1,4 @@
+#include "bet.h"
+cJSON *bet_read_json_file(char *file_name);
+void bet_parse_config_file();
+void bet_parse_notary_file();
